feat: add test example number
