+++ TestsExample/App.js
@@ -64,12 +64,12 @@ import Test1157 from './src/Test1157';
 import Test1162 from './src/Test1162';
 import Test1188 from './src/Test1188';
 import TestFreeze from './src/TestFreeze';
-import TestNew from './src/TestNew';
+import Test1214 from './src/Test1214';
 
 export default function App() {
   return (
     <ReanimatedScreenProvider>
-      <TestNew />
+      <Test42 />
     </ReanimatedScreenProvider>
   );
 }
